pyproject.toml
[project]
name = "mkdocs-macros-utils"
version = "0.1.0"
description = "MkDocs Macros plugin for creating enhanced cards and code blocks"
authors = [
    {name = "7rikazhexde", email = "33836132+7rikazhexde@users.noreply.github.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mkdocs-macros-plugin>=1.3.7,<2.0.0",
    "mkdocs-material>=9.6.1,<10.0.0",
    "requests>=2.25.0",
    "jinja2>=3.0.0",
    "pygments>=2.19.1"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
mypy = "^1.14.1"
ruff = "^0.9.4"
pre-commit = "^4.1.0"

[[tool.mypy.overrides]]
module = ["mkdocs.*","mkdocs_macros.*","jinja2.*","ruamel.*","pygments.*"]
ignore_missing_imports = true

[project.entry-points."mkdocs.plugins"]
"macros-macros-utils" = "mkdocs_macros_utils"
version = "0.1.0"
